+++ lib/features/auth/domain/entities/role_entity.dart
@@ -38,4 +38,5 @@ class RoleEntity {
       saleTypes: saleTypes ?? this.saleTypes,
     );
   }
+  
 }