+++ lib/features/point_of_sale/presentation/views/dialogs/close_po_dialog.dart
@@ -103,7 +103,7 @@ class ClosePoDialog extends ConsumerWidget {
                                         spacing: 5,
                                         children: [
                                           Text(
-                                            userData.user!.posInfo.name,
+                                            userData.user!.posInfo?.name ?? "",
                                             style: GoogleFonts.poppins(
                                               color: Colors.black.withValues(alpha: .8),
                                               fontSize: 13,
@@ -111,7 +111,7 @@ class ClosePoDialog extends ConsumerWidget {
                                             ),
                                                                 ),
                                           Text(
-                                            "(${userData.user!.posInfo.code})",
+                                            "(${userData.user!.posInfo?.code ?? ""})",
                                             style: GoogleFonts.poppins(
                                               color: Colors.grey,
                                               fontSize: 13,