+++ lib/features/point_of_sale/presentation/widgets/account_menu/header_menu/current_turn_end_date.dart
@@ -75,7 +75,7 @@ class CurrentTurnEndDate extends ConsumerWidget {
                 }
               }
 
-              ref.read(pointOfSaleProvider.notifier).closeTurn(context: context);
+              //ref.read(pointOfSaleProvider.notifier).closeTurn(context: context);
 
               ref.read(authProvider.notifier).removeCurrentTurn();