+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+### [00012] - [12/AGO/2025] - @JULIO VILLAGRANA
+- Add table management feature 
+
 ### [00011] - [11/AGO/2025] - @JULIO VILLAGRANA
 - Scanner order feature added